Primarily fixed formatting of documentation to satisfy doclet and make compile:doc succeed

+++ src/main/java/com/n1analytics/paillier/EncodedNumber.java
@@ -13,9 +13,9 @@
  */
 package com.n1analytics.paillier;
 
-import com.n1analytics.paillier.util.HashChain;
-
 import java.math.BigInteger;
+
+import com.n1analytics.paillier.util.HashChain;
 
 /**
  * A class representing encoded numbers, which enables Paillier to operate on
@@ -34,6 +34,10 @@ public final class EncodedNumber {
    * <code>context.getModulus()</code>.
    * @param context The context used to encode <code>value</code>.
    * @param value The encoded value.
+   *
+   * @param context Paillier context for this instance
+   * @param value value to encode
+   * @param exponent exponent to use
    */
   protected EncodedNumber(PaillierContext context, BigInteger value, int exponent) {
     if (context == null) {
@@ -76,7 +80,7 @@ public final class EncodedNumber {
   public boolean isValid() {
     return context.isValid(this);
   }
-    
+
     /* TODO what is an appropriate value for invalid numbers?
     public int signum() {
     	if(value.equals(BigInteger.ZERO))
@@ -195,7 +199,7 @@ public final class EncodedNumber {
     //      corresponding positive value.
     return subtract(context.encode(other));
   }
-    
+
     /*
     public EncryptedNumber subtractUnobfuscated(EncryptedNumber other) {
         // TODO be careful not to use negate() otherwise it won't work when
@@ -232,15 +236,15 @@ public final class EncodedNumber {
     public EncodedNumber multiplicativeInverse() {
     	return context.multiplicativeInverse(this);
     }
-    
+
     public EncodedNumber divide(EncodedNumber other) {
     	return context.divide(this, other);
     }
-    
+
     public EncodedNumber divide(Number other) {
     	return divide(context.encode(other));
     }
-    
+
     public EncodedNumber divide(BigInteger other) {
     	return divide(context.encode(other));
     }
@@ -251,7 +255,7 @@ public final class EncodedNumber {
   }
 
   public EncodedNumber divide(long other) {
-    return multiply(context.encode(1.0 / (double) other)); // TODO unhack
+    return multiply(context.encode(1.0 / other)); // TODO unhack
   }
 
   @Override
