+++ src/main/java/com/n1analytics/paillier/EncryptedNumber.java
@@ -13,9 +13,9 @@
  */
 package com.n1analytics.paillier;
 
-import com.n1analytics.paillier.util.HashChain;
-
 import java.math.BigInteger;
+
+import com.n1analytics.paillier.util.HashChain;
 
 /**
  * Immutable class representing encrypted number and arithmetic operations that can be computed on the encrypted number.
@@ -107,7 +107,7 @@ public final class EncryptedNumber {
    * @param context PaillierContext used to encrypt this encrypted number.
    * @param ciphertext the encrypted representation of the encoded number.
    * @param exponent the exponent of the ciphertext.
-   * @prama isSafe set to true if cypertext is obfuscated
+   * @param isSafe set to true if cypertext is obfuscated
    */
   public EncryptedNumber(PaillierContext context, BigInteger ciphertext, int exponent,
                          boolean isSafe) {
@@ -261,11 +261,11 @@ public final class EncryptedNumber {
     public EncryptedNumber divide(EncodedNumber other) {
     	return context.divide(this, other);
     }
-    
+
     public EncryptedNumber divide(Number other) {
     	return divide(context.encode(other));
     }
-    
+
     public EncryptedNumber divide(BigInteger other) {
     	return divide(context.encode(other));
     }
@@ -276,7 +276,7 @@ public final class EncryptedNumber {
   }
 
   public EncryptedNumber divide(long other) {
-    return multiply(context.encode(1.0 / (double) other)); // TODO unhack
+    return multiply(context.encode(1.0 / other)); // TODO unhack
   }
 
   public void serialize(Serializer serializer) {
